Major refactoring, new video representation using xarray and dask.
This commit marks the transition from a pure Numpy architecture to a more
scalable implementation based on Dask.  This allows for chunk-wise processing,
lazy IO, and a lot of other conveniences.

Special care is taken such that videos processed with this library have a
canonical representation.  The function toolsandogh.io.verify_video asserts that a
video adheres to this representation.

Structural changes:

- The toolsandogh module has been moved to src/toolsandogh

- The iscat_analysis.py script has been moved to scripts/iscat_analysis.py